commitV2.4b: right_padding_token_emb 三行 dtype 修复——消除 framesample 短样本的 f64 提升与双 dtype 路径
本 commit 是 dtype 统一计划的**唯一功能改动**，也是**唯一的 revert 对象**。验证工具
在 V2.4a `f2e7348` 已先行落地，修复前取证（P3）已在那个 clean HEAD 上完成，故本
commit 之后两侧产物可直接对拍。

一、用户指令原话

  「/nfs/turbo/coe-chaijy-unreplicated/hongzefu/robomme_policy_learning_MotionJEPA/
    v1-dtype-unify-plan.md  实现 有任何疑问现在问 不要自己做决策」

与本 commit 直接相关的两条裁定：
  - 「我只关注 framsample+context！」→ `left_padding_token_emb`（recurrent 路径）
    有完全同型的三处无 dtype `np.zeros`，本轮**不动、也不加注释**；
  - 修复无运行时开关，A/B 是跨 commit 的，diff 面必须精确到三行。

二、改动内容（src/mme_vla_suite/shared/data_utils.py，唯一文件）

`right_padding_token_emb` 短样本分支的三个填充块各加 dtype 参数：

    np.zeros((max_size - sampled_img_emb.shape[0],   *sampled_img_emb.shape[1:]),   dtype=sampled_img_emb.dtype)
    np.zeros((max_size - sampled_pos_emb.shape[0],   *sampled_pos_emb.shape[1:]),   dtype=sampled_pos_emb.dtype)
    np.zeros((max_size - sampled_state_emb.shape[0], *sampled_state_emb.shape[1:]), dtype=sampled_state_emb.dtype)

零改动：mask 分支（已显式 `dtype=np.bool_`）、满长切片分支、`left_padding_token_emb`、
以及其他一切文件。`git diff` 三个 hunk，各 +2/−1。

三、为什么这不是引入新行为

numpy 的 `np.zeros` 默认 float64，`np.concatenate` 按提升规则把整条序列一起抬上去。
而模型第一层 `nnx.Linear` 显式 `dtype=bfloat16`，flax `promote_dtype` 在任何算术前
把输入统一转 bf16；bf16→f32→f64 是精确升位（全部 65,536 个 bf16 位型验证过往返
无损）。所以三种交付（bf16 / f32 / f64）进 `pos_proj` / `encoder_static` 的实际张量
逐位相同——**修复不是改变行为，而是把现状 1.6% 满长 batch 的行为推广到 100%**。

`max_size = budget 512 // (token_per_image 16 × num_views 1) = 32`，
`even_sampling_indices` 在 `step_idx < max_size` 时返回 `range(step_idx+1)`，
所以短样本边界是 **`step_idx ≤ 30`**（原计划正文口语写的「t < 31」已在文档修订中订正）。

四、修复前现状的实测证据（P3，run v1-dtype-p3-dump-pre，起跑 commit f2e7348）

逐样本（2,600 个定点样本，裸 RoboMMEDataset）：
  短样本 1,068 个 → static_image_emb 100% float64
  满长样本 1,532 个 → static_image_emb 100% bfloat16
  零例外，双路径完整成立。

batch 级（200 个定点 batch，完整 transform_dataset + _collate_fn）：
  allfull （全满长） static_image_emb bfloat16 ×50 / static_pos_emb float32 ×50
  allshort（全短样本）static_image_emb float64  ×50 / static_pos_emb float64 ×50
  mixed1 （1 短 + 7 满长）同 allshort，全 float64 ×50
  random （随机组成） static_image_emb bfloat16 ×28 + float64 ×22
                      static_pos_emb  float32  ×28 + float64 ×22
  static_state_emb 四档恒 float64（经 _normalize_state，norm stats 为 f64）

random 档 28/22 的分裂是「XLA 被迫编译两份产物」最直接的证据——**同一档随机 batch
里 dtype 就在摆动**。22/50 = 44% 含短样本，与理论值吻合（随机池含 6.8% 短样本，
1 − 0.932^8 = 43.0%）。

五、修复后的验证（本 commit）

`JAX_PLATFORMS=cpu uv run pytest scripts/dtype-unify/ -q` → **14 passed**（42.4s）。
修复前同一命令是 7 passed / 7 skipped——那 7 条 skip 正是纯函数 dtype 断言，测试
用「探测修复是否已落地」自动切换：未落地时 skip 并说明原因，落地后自动转为真断言，
不需要改测试文件。本 commit 使它们全部转正并通过，覆盖 t ∈ {1,2,3,30,31,32,33} 七档：
  - 三键输出 dtype 跟随各自输入（img bf16 / pos f32 / state f32）；
  - 非填充区逐字节不变；
  - 填充区全零；
  - mask 恒 bool、有效区全 True、填充区全 False。

这组纯函数测试是 `static_state_emb` 那一处修复的**唯一**有效验证——它的交付键经
`_normalize_state`（norm stats q01/q99 为 f64）恒为 f64，修复前后不变，在交付键和
梯度上都不可观测。

六、影响面与验收范围（分开声明）

实际影响面：全仓库唯一调用点是 `MemoryBuffer._prepare_frame_sampling`，该函数不区分
context / modulation / expert 变体，且同时被训练侧（training/dataset.py）与在线评估侧
（policies/policy.py 的 history 准备路径）调用——修复会一并改变全部 frame_sampling
变体与在线评估的 padding dtype。recurrent 路径走 `left_padding_token_emb`（未修），
symbolic / token_dropping 零波及。

验收范围：本计划两块验证只覆盖 `perceptual-framesamp-context` 训练路径。其余受影响
路径（在线评估、modulation / expert 变体）**不在验收范围**——修复同为「补 dtype 参数、
填充值恒 0」的精确升位消除，风险同质，但不得据此宣称已验证；函数级证据由上面那组
纯函数位型测试提供。

已知未修：`left_padding_token_emb` 有完全同型的三处无 dtype `np.zeros`（零块拼在
concat 列表第一项、切片为 `[-max_recur_steps:]`）。按用户裁定「只关注
framesample+context」本轮不动，此处记录以免日后 recurrent 路径重现同一缺陷时重复调查。

七、当前状态与下一步

链路状态：修复已落地，纯函数级等价性已证；数据集级与训练级等价性待 P4–P6 验收。
P3 的 A 侧产物齐备（7.1 GB dump + 33 GB 梯度数组在 /data/hongzefu/v1-baselines/
dtype-p5-grad-pre/ + 80 MB A 侧 batch fixture），可直接作为对拍一侧。

下一步：
  P4 v1-dtype-p4-cmp     修复后 dump + 第一块三层对拍（判据 COMPARE_DTYPE=PASS）
  P5 v1-dtype-p5-grad    B 侧单步梯度 + 与 P3 的 A 侧逐元素对拍（COMPARE_GRAD=PASS）
  P6 v1-dtype-ab-post-r1 1000 步 G1 vs G0b r1（bitwise，失败时量化兜底）
  P7 v1-g1-speed         1000 步性能 run vs v1-g0-speed-r2

任何一块最终不过 → revert 本 commit（工具与失败证据不回滚），计划关闭。

+++ src/mme_vla_suite/shared/data_utils.py
@@ -28,7 +28,8 @@ def right_padding_token_emb(
                     (
                         max_size - sampled_img_emb.shape[0],
                         *sampled_img_emb.shape[1:],
-                    )
+                    ),
+                    dtype=sampled_img_emb.dtype,
                 ),
             ],
             axis=0,
@@ -40,7 +41,8 @@ def right_padding_token_emb(
                     (
                         max_size - sampled_pos_emb.shape[0],
                         *sampled_pos_emb.shape[1:],
-                    )
+                    ),
+                    dtype=sampled_pos_emb.dtype,
                 ),
             ],
             axis=0,
@@ -52,7 +54,8 @@ def right_padding_token_emb(
                     (
                         max_size - sampled_state_emb.shape[0],
                         *sampled_state_emb.shape[1:],
-                    )
+                    ),
+                    dtype=sampled_state_emb.dtype,
                 ),
             ],
             axis=0,
